requirements.txt
re
requests
os
fpdf
pillow